requirements.txt
fastapi
uvicorn
numpy
Pillow
ultralytics
google-generativeai
PyYAML
python-multipart
httpx
fuzzywuzzy[speedup]
